fix tranp , fix leave

+++ app/src/main/java/com/app/ecarepro/data/UserRepositoryImpl.kt
@@ -333,7 +333,7 @@ class UserRepositoryImpl @Inject constructor(
         leaveID: Int,
         fromDate: String,
         tillDate: String,
-        duration: Int,
+        duration: Double,
         halfdayDTL: List<HalfdayDTL>,
         reason: String,
         attachment: String,
