+++ app/src/main/java/com/app/ecarepro/data/repository/UserRepository.kt
@@ -217,7 +217,7 @@ interface UserRepository {
         leaveID: Int,
         fromDate: String,
         tillDate: String,
-        duration: Int,
+        duration: Double,
         halfdayDTL: List<HalfdayDTL>,
         reason: String,
         attachment: String,
